refactor(regression): Separate module for regression structs for all tools
